Added BattleViewer to GameView and hooked it up to message button. BattleViewer now has individual "GridToken" nodes that contain state like their current damage and such Renamed all String to string All Rules() creations use the same seed for consistency, except when you actually create a game from the MainMenu. Split Beam/Torpedo fire into separate records Implemented FireTorpedo Torpedos make one record per token, rather than per missile Fleets get their BattlePlan from the Player's BattlePlans. Starbases have +1 RangeBonus BattleComputers impact Torpedo Accuracy

+++ src/Core/Models/Tech/TechHullComponent.cs
@@ -13,7 +13,13 @@ namespace CraigStars
         public int Shield { get; set; }
         public int Cloak { get; set; }
         public bool CloakUnarmedOnly { get; set; }
-        public int TorpedoBonus { get; set; }
+        
+        /// <summary>
+        /// This cumulative bonus decreases innaccuracy of torpedos on the ship
+        /// i.e. 75% accurate torpedo with two .3f torpedo bonus
+        /// is 100 - ((100 - 75) x .7 * .7) = 88% accuracy
+        /// </summary>
+        public float TorpedoBonus { get; set; }
         public int InitiativeBonus { get; set; }
         public int TorpedoJamming { get; set; }
         public int ReduceMovement { get; set; }
